Refactor: html 로직 분리
