+++ js/main/newsstnad.js
@@ -1,6 +1,6 @@
 import { shuffle } from '../utils/utils.js';
 import { getNewsData } from '../core/apis.js';
-import { html } from '../core/createElement.js';
+import { paintNewsstand } from '../components/newStandGrid.js';
 
 let SELECTED_PAGE = 0;
 const $ul = document.querySelector('.newsstand-area—six-col-list');
@@ -10,35 +10,22 @@ const $leftBtn = document.querySelector('.newsstand--left-btn');
 async function initNewsStand() {
   const newsData = await getNewsData();
   const datas = newsDataPaser(shuffle(newsData));
-  paintNewsstand(datas);
+  paintNewsstand(datas, SELECTED_PAGE);
   pagination(datas);
-}
-
-function paintNewsstand(datas) {
-  let htmls = '';
-  for (let idx = SELECTED_PAGE * 24; idx < SELECTED_PAGE * 24 + 24; idx++) {
-    const [name, src] = datas[idx];
-    htmls += html`
-      <li class="newsstand—subscrtion-box">
-        <img src="${src}" alt="${name}" />
-      </li>
-    `;
-  }
-  $ul.insertAdjacentHTML('beforeend', htmls);
 }
 
 function pagination(datas) {
   const handleRightBtn = () => {
     $ul.innerHTML = '';
     ++SELECTED_PAGE;
-    paintNewsstand(datas);
+    paintNewsstand(datas, SELECTED_PAGE);
     isBtnDisabled();
   };
 
   const handleLeftBtn = () => {
     $ul.innerHTML = '';
     --SELECTED_PAGE;
-    paintNewsstand(datas);
+    paintNewsstand(datas, SELECTED_PAGE);
     isBtnDisabled();
   };
