+++ js/main/rollingSubView.js
@@ -1,5 +1,5 @@
 import { getRollingData } from '../core/apis.js';
-import { html } from '../core/createElement.js';
+import { createSubView } from '../components/subView.js';
 
 let RollingData = [];
 let Lidx = 1;
@@ -11,26 +11,6 @@ async function initSubView() {
   createSubView(datas);
   RollingData = getSubViewData(datas);
   startAnimation();
-}
-
-function createSubView(datas) {
-  let htmls = '';
-  datas.map((data) => {
-    const [name] = data;
-    htmls += html`
-      <div class="main__content">
-        <div class="main-—corp-name">${name}</div>
-        <div class="main—-news-area">
-          <div class="main—-news-box animate-subview-auto-rolling">
-            <div class="main—-news-subview"></div>
-            <div class="main—-news-subview"></div>
-          </div>
-        </div>
-      </div>
-    `;
-  });
-
-  document.querySelector('.main__subview').insertAdjacentHTML('beforeend', htmls);
 }
 
 function startAnimation() {
@@ -120,6 +100,7 @@ function changeRightNewsData(rightData) {
   return data;
 }
 
+// 데이터 계산
 function getSubViewData(datas) {
   datas = datas
     .map((d) => [...d[1]])
